Reran entity framework to add a property to customer
