+++ Src/CustomerManagement/View/UserControls/CustomerDataGrid.xaml
@@ -11,6 +11,7 @@
         <DataGrid AutoGenerateColumns="False" Name="CustomerTable" RowEditEnding="CustomerTable_RowEditEnding">
             <DataGrid.Columns>
                 <DataGridTextColumn Header="Customer ID" Binding="{Binding Id, UpdateSourceTrigger=PropertyChanged}" />
+                <DataGridTextColumn Header="Sage Ref" Binding="{Binding SageRef, UpdateSourceTrigger=PropertyChanged}" />
                 <DataGridTextColumn Header="Account Name" Binding="{Binding CompanyName, UpdateSourceTrigger=PropertyChanged}" />
                 <DataGridTextColumn Header="Business Contact" Binding="{Binding BusinessContact, UpdateSourceTrigger=PropertyChanged}" />
                 <DataGridTextColumn Header="Email" Binding="{Binding EmailAddress, UpdateSourceTrigger=PropertyChanged}" />
